# be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Thăng Long Chè Việt — Storefront API",
        "description": "Public REST API for Nuxt storefront: products, blog, cart, orders, booking, contact. Admin UI uses Filament at /admin.",
        "contact": {
            "email": "admin@thanglongcheviet.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/storefront/coupons/validate": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Validate coupon code",
                "description": "Check whether a coupon is valid and compute discount for the current cart total.",
                "operationId": "validateCoupon",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/LangQuery"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "cart_total"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SUMMER10"
                                    },
                                    "cart_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "valid": {
                                                    "type": "boolean"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "value": {
                                                    "type": "number",
                                                    "nullable": true
                                                },
                                                "discount": {
                                                    "type": "number",
                                                    "nullable": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/storefront/products/{slug}/reviews": {
            "get": {
                "tags": [
                    "Products",
                    "Reviews"
                ],
                "summary": "List approved product reviews",
                "operationId": "getProductReviews",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/LangQuery"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review list"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/storefront/booking": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Submit salon/spa booking request",
                "description": "Creates a booking inquiry for salon services. Stored in admin under Contact Inquiries.",
                "operationId": "submitBooking",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/LangQuery"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0901234567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "service": {
                                        "description": "Service slug or name",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "staff_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "preferred_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01",
                                        "nullable": true
                                    },
                                    "preferred_time": {
                                        "type": "string",
                                        "example": "10:00",
                                        "nullable": true
                                    },
                                    "note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking submitted"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
        "/api/storefront/booking/availability": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Available booking time slots",
                "description": "Returns 30-minute slots for a given date based on salon hours and existing bookings.",
                "operationId": "getBookingAvailability",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-07-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot list"
                    },
                    "422": {
                        "description": "Invalid date"
                    }
                }
            }
        },
...
        "/api/storefront/checkout": {
            "post": {
                "tags": [
                    "Cart",
                    "Orders"
                ],
                "summary": "Place order from cart",
                "operationId": "checkout",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/SessionIdHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "coupon_code": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created"
                    },
                    "400": {
                        "description": "Empty cart or invalid coupon"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "LangQuery": {
                "name": "lang",
                "in": "query",
                "description": "Locale for translatable content (vi | en)",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "vi",
                    "enum": [
                        "vi",
                        "en"
                    ]
                }
            },
            "SessionIdHeader": {
                "name": "X-Session-ID",
                "in": "header",
                "description": "Anonymous cart session UUID (required for cart/checkout)",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
...
        },
        "securitySchemes": {
            "customerBearer": {
                "type": "http",
                "description": "Customer API token (not admin token)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}